fixed test cases forwarding
issue 1 was decode get_buffer_data() was not being passed data.read_register_# when buffer 5 or more dependency, so it was using default, which is incorrect.

issue 2 was that at clock cycle 82, with lw instruction, the jump target was the same as pc+4, which the flush check evaluated as true, because the comparison was new_pc==jump_target, which was true, even though there was no flush. now it is the correct condition, which is jump_taken boolean.

final issue was that it took >100 instruction to finish (with pipeline) so max_instructions of 100 was preventing it from finishing. set it to 128 now.